requirements.txt
openai
google-genai
google-api-python-client
discord
PyYAML